Moved to Navbar, Added toolbar
