+++ client/src/components/Editor/DraftEditor.js
@@ -8,10 +8,8 @@ import "prismjs/themes/prism-okaidia.css";
 import "./css/draft.css";
 // Material-UI imports
 import { makeStyles } from "@material-ui/core/styles";
-import IconButton from "@material-ui/core/IconButton";
-import Icon from "@material-ui/core/Icon";
-import Grid from "@material-ui/core/Grid";
-import Box from "@material-ui/core/Box";
+import Toolbar from "./Toolbar";
+
 // Prism language support:
 import {
   go,
@@ -26,6 +24,18 @@ import {
   languagesGrammar,
 } from "./languages";
 
+console.log(
+  "Prism langs imported: ",
+  go,
+  clike,
+  ruby,
+  php,
+  python,
+  c,
+  cpp,
+  csharp,
+  java
+);
 const useStyles = makeStyles((theme) => ({
   editor: {
     border: "1px solid #ddd",
@@ -39,12 +49,7 @@ const useStyles = makeStyles((theme) => ({
 const MyEditor = ({ language, makeSubmit, onSubmit, hasContent }) => {
   const classes = useStyles();
 
-  let mappedLanguage;
-  if (language === "") {
-    mappedLanguage = null;
-  } else {
-    mappedLanguage = languagesGrammar[language];
-  }
+  const mappedLanguage = languagesGrammar[language] || null;
 
   const decorator = new PrismDecorator({
     prism: Prism,
@@ -84,7 +89,6 @@ const MyEditor = ({ language, makeSubmit, onSubmit, hasContent }) => {
     );
   }, [editorState, mappedLanguage]);
 
-  // Focus editor (we will chat using the same editor, right?)
   const editor = useRef(null);
   const focusEditor = () => {
     editor.current.focus();
@@ -155,70 +159,13 @@ const MyEditor = ({ language, makeSubmit, onSubmit, hasContent }) => {
     }
   }, [makeSubmit]);
 
-  // Rendering control buttons
-  const ToolButton = ({ format, icon }) => {
-    const currentStyle = editorState.getCurrentInlineStyle();
-    return (
-      <IconButton
-        size="medium"
-        color={currentStyle.has(format) ? "secondary" : "default"}
-        onMouseDown={(event) => {
-          event.preventDefault();
-          let nextState = RichUtils.toggleInlineStyle(editorState, format);
-          setEditorState(nextState);
-        }}
-      >
-        <Icon>{icon}</Icon>
-      </IconButton>
-    );
-  };
-
-  const BlockButton = ({ format, icon }) => {
-    const selection = editorState.getSelection();
-    const blockType = editorState
-      .getCurrentContent()
-      .getBlockForKey(selection.getStartKey())
-      .getType();
-    return (
-      <IconButton
-        size="medium"
-        color={format === blockType ? "secondary" : "default"}
-        onMouseDown={(event) => {
-          event.preventDefault();
-          let nextState = RichUtils.toggleBlockType(editorState, format);
-          setEditorState(nextState);
-        }}
-      >
-        <Icon>{icon}</Icon>
-      </IconButton>
-    );
-  };
-
   return (
     <div className={classes.editor}>
-      <div className={classes.controls}>
-        <Grid container justify="flex-start">
-          <Grid item xs={8} sm="auto" md="auto">
-            <ToolButton format="BOLD" icon="format_bold" />
-            <ToolButton format="ITALIC" icon="format_italic" />
-            <ToolButton format="UNDERLINE" icon="format_underlined" />
-            <BlockButton format="code-block" icon="code" />
-          </Grid>
-          <Grid item xs={0} sm={5} md>
-            <Box display={{ xs: "none", sm: "block" }}>
-              <BlockButton format="blockquote" icon="format_quote" />
-              <BlockButton
-                format="ordered-list-item"
-                icon="format_list_numbered"
-              />
-              <BlockButton
-                format="unordered-list-item"
-                icon="format_list_bulleted"
-              />
-            </Box>
-          </Grid>
-        </Grid>
-      </div>
+      <Toolbar
+        editorState={editorState}
+        RichUtils={RichUtils}
+        setEditorState={setEditorState}
+      />
       <Editor
         editorState={editorState}
         onChange={onChange}
